Prohibit caches of views of dimension records.
This avoids a problem with dimension record cache consistency on was
introduced in the cache rework a few commits back, when we started
considering the cache more trustworthy.

Unfortunately there's no cost-free fix: that cache rework fixed other
bugs with cache consistency, so we can't just drop it.  The best
solution seems to be admitting that expandDataIds cannot be _relied_
upon to test for dimension record existence; this has always been the
case in general, but it may have been impossible to trigger via the
default dimension configuration before.

See code comments for additional rationale.