Use context manager to enforce discipline over cursor lifetimes.
This addresses a long-standing TODO comment in Database.query (dating
back to the introduction Session and our first real attempt to reduce
connection contention).  In some cases this might seem to be
transforming lazy iteration over database results into aggressive
fetching, via calls to fetchall(), but for at least PostgreSQL it seems
we were already doing _some_ aggressive fetching before without
realizing it, see e.g.

https://docs.sqlalchemy.org/en/14/core/connections.html#using-server-side-cursors-a-k-a-stream-results

So what we're really doing in the new fetchalls() here is invoking
SQLAlchemy's client-side row processing more aggressively.   I don't
think that's a big change, and in exchange we can now guarantee that
we never return iterators to user code that are responsible for closing
a connection when the user is "done" with them - a problematic
definition for any garbage-collected entity, but a particular problem
for iterators that might not ever be exhausted.

This also include some typing adjustments for SQLAlchemy objects.
Those aren't actually checked yet, since typing support in SQLAlchemy
won't be out until it's 2.0 release.