Reuse connections more at butler startup.
Each of the gazillion small queries we run a butler startup was
previously grabbing a new connection and then returning it.  That was
probably wasn't too bad, since it should have just been getting them
from the pool (for PostgreSQL) or opening a local file (for SQLite),
but this should still be better in both cases.

It should also remove a lot of rollbacks, since SQLAlchemy emits those
whenever connections are returned to the pool.  SQLAlchemy's docs
say those shouldn't matter in terms of performance, but they're still
noisy.