Improve edge-case handling in Registry.findDatasets.
This was originally supposed to just add a test case for DM-37504, in
which attempting to search both a CALIBRATION collection and a
non-CALIBRATION collection for the same dataset type (with collection-
summary optimization not in play) triggered a bug in daf_relation.

But while writing the test I also discovered an unhelpful error message
when trying perform that search without the needed timespan argument,
so I've updated the logic in findDataset to ensure the better error
message already there is emitted in more (now, hopefully all) cases.