Add test for the get_uri method
This is not going to be a production method but is there
as a demonstration. To make it work we had to implement
refresh in the server but the long term usefulness of this
needs thought. What happens if a dataset type or collection
is added to the server butler? How do the cached butler objects
in the server know they all need to update?