Add new test script for checking query relation trees.
This provides a new approach to testing registry queries that won't
require as much work setting up appropriate test data and inspecting it
to independently determine expected query results, and I'm looking
forward to adding more tests here on future tickets; so far this just
focuses on the spatial constraint queries that we apparently had
particularly poor test coverage previously, leading to lots of breakage
on DM-31725.

Some of these tests are expected to fail right now, but will be fixed
shortly.