Remove currently-unusable temporal constraint logic in queries.
There's currently no high-level way to populate these arguments to our
lower-level query classes; the only way one can currently do temporal
constraints is to write them out in the 'where' argument, and that
doesn't use any of this.

I've made this a separate commit because I think we may want to revert
it after we've got a public interface (possibly on Butler rather than
Registry) that's designed to support all of the things the daf_relation
query system could now do.  But for now it's just dead weight.