Add obscore manager property to Registry interface.
Some clients, e.g. DefineVisitsTask, will need an access to ObsCore update
methods. I do not want to expose those methods in Registry interface because
they are only relevant for a small set of clients and a small number of repos.
Instead, an `obsCoreTableManager` property is added which will be `None` if
obscore is not configured for a repo.