Deprecate non-configurable access to dimension packers.
This also moves test coverage for what will remain (here) of the
dimension packer system to test_dimensions.py, relying on a new
simpler concrete test packer.  All of the production concrete packers
will be defined in downstream packages going forward, and the
machinery for constructing them from a DimensionUniverse is being
deprecated, so all that remains to be tested is the DimensionPacker
ABC itself.