Small fix for Butler.makeRepo method (DM-38814)
In case when we need to drop obscore config key from registry before writing
out butler.yaml file, we make a copy of the config, modify it and write that
copy. Writing it out modifies its `configFile` attribute and we need to
propagate its new value back to the original config. I have added new unit
test to `pipe_base` that uses obscore and found that it needs this change.