Make StorageClass.converters_by_type private and fix annotation.
This is only used by other StorageClass methods right now, and locking
it down as private will make it possible to implement MockStorageClass
(in pipe_base) by making that method raise and reimplementing the other
StorageClass methods that call it.

I've made it into a _get method instead of leaving it as a property
since that sidesteps the question of what to call it (as there's
already a private attribute) and for an internal interface we don't
need the prettiness of attribute-style access.