Make StorageClass methods/properties return Mapping, not dict.
It looks like these are only properties to discourage assignment, but
that's not worth much if even type-checked can modify the returned
thing in-place.

But the impetus for doing this now is that Mapping is covariant, and
hence a StorageClass subclass (i.e. MockStorageClass, being added to
pipe_base) can return a mapping whose values are guaranteed to be
MockStorageClass, not just StorageClass.