Update Butler.put method to pass unit test with resolved ref.
Butler.put now raises an exception for duplicate attempt to write the same
resolved ref. There is still a difference in put() behavior between resolved
and unresolved (DatasetType+DataId) cases when writing a dataset that has
registry records but no artifacts. This is now reflected in the unit test,
and I am not sure what the correct behavior would be in this case.