Record read storage class before casting ref in Datastore.get.
This was a minor bug introduced on DM-39198 that caused Datastore.get
to pass the wrong readStorageClass to the formatter in trust mode.
I'm not sure if this led to storage class conversions on get always not
returning the right storage class (and this going unnoticed because
the only conversions in play have Python types with intentionally
similar interfaces, like PropertySet and TaskMetadata), or if this
error was usually corrected later in Datastore.get in a check on the
pytype returned.  I noticed it when using the new mock storage classes,
for which that later check on pytype does nothing, since all mock
storage classes have the same pytype.