Conform storage classes in Datastore even when not in trust mode.
Direct execution against a full butler (i.e no execution butler or
quantum-backed butler) revealed that `Butler.put` was still writing
with the wrong storage class (but at least recording that wrong storage
class into the Datastore records), and hence Datastore's apparent
assumption that Butler would take care of passing it the right
DatasetRef was incorrect.

I'm not sure what logic exists still in Butler for conforming storage
classes to the registry definition, but this is at least another step
(and maybe the last step) to making Datastore not care.  I'm not sure
that's the right division of responsibilities long-term, but I'm not
sure it isn't, and I think it's the only viable one as long as
ctrl_mpexec or anything else calls Butler.datastore methods directly.