Enable repository index in QBB and `butler config-dump` (DM-39429)
Extended `ButlerRepoIndex.get_repo_uri` method with a `default` parameter,
if label is not found in an index then default is returned. Updated QBB and
`configDump` to use that method to try to resolve repo in the index. Added
a unit test for QBB, `config-dump` was tested manually. Also changed `Butler`
use of that method to pass a default to simplify its logic.