Mark Butler._exists_many as private
We are not entirely sure that this is the right API given the
possibilities that refs can be given that were crafted externally
but still look like they share the same ID (storage class overrides
notwithstanding). The API would be cleaner if we used iterable[UUID]
instead but that needs some thought. For now mark it private but
still test it.