Change file formatter to support v1 and v2 pydantic constructors
The pydantic models that use the v1 compatibility class work
fine with model_validate but there are some v1 models that do
not inherit from that and also some classes like ScarletModelData
that emulate the pydantic v1 methods but are not using pydantic.

Therefore we need to still check parse_obj.