Move StrictInt ahead of StrictFloat in model
In newer pydantic this converts a 0 to 0.0 when serializing
to Json and this breaks when reading a record back that is
meant to be an integer. This changes in pydantic 2.2 whereas
2.1 behaved like 1.0.

This happens because the dynamic subclasses apply dynamic constraints
based on the specific record data model so they know that something
must be an integer.