Improve error messages for dataset import conflicts.
Just had to debug some of these, and found the integer primary keys
embedded in the old messages hard to use.  Unfortunately I couldn't get
rid of all of them: this code has no access to the general mapping
from dataset type ID to dataset type name.  But I think that's the
least likely field to be causing the conflict here, so it's not a huge
loss.