Remove datastore records from serialized DatasetRef (DM-41226)
This removes the piece of code for (de-)serializing datastore records in
DatasetRef that was added on DM-40053. We worry that serialized QuantumGraph
could grow in size, though presently we do not fill datastore records when
we make refs for quantum graph. In the future the whole thing will be
replaced by a more efficient data structures, so this serialization will
likely disappear entirely.