Add DimensionGroup and deprecate DimensionGraph.
This adds a lot of warnings; DimensionGraph usage will be migrated in
later commits.

Due to those warnings, one test in test_obscore.py has been
temporarily disabled, because it counts warnings, and I couldn't
figure out how to make it only count the kind of warning it's actually
looking for.