Switch to DimensionGroup in DatasetType and LookupKey.
DatasetType.dimensions needs to continue to return DimensionGraph
during the deprecation period, since that's the name we want to use
long-term (unlike e.g. DataCoordinate.graph).  That means we rely on
the fact that we've also deprecated all of the things DimensionGraph
can do that DimensionGroup can't do (like iteration), and we rely on
those warnings instead of making DatasetType.dimensions itself warn.
That should let a lot of usage just blindly pass
DatasetType.dimensions to something else and not care about what type
it is throughout the deprecation period.