Add `Butler._query` context manager (DM-41761)
This new method defines an interface for making complex queries
on repository database. This initial commit includes only three
convenience methods that are a close match  to the query methods that
already exist in Registry. The implementation of that interface
for DirectButler just forwards calls to SqlRegistry.

RemoteButler does not implement those methods yet, we will need
to define representation for queries and results that client exchanges
server. This should become clearer as we add more methods to `Query`.