Prototype partial(?) replacements relation-tree objects.
The current daf_relation tree in our Query objects is too contaminated
with implementation details to serve us well in the RemoteButler
client, and we could only partially mitigate that by defining a new
daf_relation Engine.  The fundamental issue is that daf_relation
objects expect to know exactly which columns they have at any given
time, and that's at odds with user expectations that columns
"magically" appear whenever they're requested (e.g. referenced by a
`where` string), and that this joins in the tables that provide them
as-needed.

The two new files here *heavily* duplicate stuff in daf_relation, and
in addition to being more vague about what the columns are, they're
simpler for two key reasons:

- They're just data, with no logic for maintaining invariants,
  contructing trees, serialization, or anything else.  This will have
  to change as we actually start to use them.

- They fully enumerate the kinds of expressions and operations we care
  about in the butler query system rather than trying to define
  abstract versions of those upstream of daf_butler that could be
  specialized in daf_butler.  I had not appreciated how much of a
  simplification this could be when writing daf_relation as a separate
  package, and if it holds up it may suggest that the right way to
  resolve the duplication is to rip a lot of stuff out of daf_relation.
  serialization.