Avoid using size() on remote ResourcePaths
Move the resource size check when retrieving artifacts so it uses the downloaded data size, instead of checking the size separately.

HEAD is not supported on S3 URLs presigned for GET, so HttpResourcePath.size() does not work for them.  This is also a slight optimization, since it saves an unnecessary network round trip.