Start replacing ddl.FieldSpec with something less SQL-focused.
When we convert the dimensions.yaml config into a DimensionUniverse,
we've been converting the rather limited set of types supported in the
configuration into just a few of the very specific types used by
SQLAlchemy (in the ddl.FieldSpec objects), and that makes it hard to
write a mapping from the Python form of that information to some other
schema system (like Pydantic or Arrow).

The new ColumnSpec objects map more directly to what's in the config
(they're Pydantic models; this is one tiny step towards making
dimensions.yaml Pydantic-validated), and while right now they live
alongside the ddl objects everywhere (there are some duplicative
methods on the DimensionElement classes in particular), eventually I'd
like to replace the ddl stuff entirely with direct usage of SQLAlchemy
Table and Column objects.  Fully removing the ddl stuff will require
an RFC as well, since they leak into the public interface a bit via
RecordClass.fields as well as the DimensionElements.