Decouple Butler constructor params from subclass
Decoupled the set of parameter passed to Butler() from those passed to its subclasses.  RemoteButler construction will be changing to require injection of some long-lived objects, so it needs a different constructor signature than the original Butler constructor.

There are several parameters to Butler like "collections" and "run" which control the behavior of a Butler instance and are common to both DirectButler and RemoteButler.  Upcoming changes to Butler construction will require these to be passed around through multiple functions.  To avoid repeating them they are pulled out to a dataclass object.