Add a factory for instantiating RemoteButlers
To provide a place for storing state that is shared between multiple instances of RemoteButler, add a RemoteButlerFactory class that can be used to get instances of RemoteButler.

This also simplifies the constructor logic for both Butler types by making Butler.from_config() the only place responsible for knowing about the configuration search logic.