Split Butler "copy constructor" into new function
The DirectButler constructor used to have two forms that were logically separate but combined into one interface:
1. Standard new instance creation
2. Copy construction

This became more awkward when part of the construction process was moved to Butler.from_config, because 'config' and 'butler' are mutually exclusive options but we always need to instantiate a config to find out which class to load.

Moved the copy construction form into a separate instance method _clone().  This will likely become public at some point in the near future.