Drop unnecessary default for dummy column in no-column temp tables.
This is an extremely edgy edge case: when we materialize a query
with no columns, we have to add a dummy column to the temp table, just
as we do to the query.  But for some reason we were giving that column
a default, and those behave strangely in SQLAlchemy (they don't go
through the usual conversion logic), and it didn't like that we were
giving a bool column a bool default value.  Easy fix is to just not
give it a default value.

The test that was supposed to exercise this code was being thwarted by
some aggressive query simplification; we could tell at query-build time
that the test query would have exactly one row, and that triggered a
different code path that skipped the temporary table entirely.