Clone inner objects when cloning DirectButler
When cloning a DirectButler instance, clone all of its inner objects instead of just copying references to the top-level properties.  This makes it safe to use the cloned objects concurrently from different threads.  This is necessary for the Butler server and other services using DirectButler as a backend, since they have a separate thread for each incoming request.