Read through DatasetType cache on a miss
Previously, if a DatasetType was not found in the cache during a call to DatasetRecordStorageManager.find(), it would return None.  Instead, we now go to the DB and fetch the data.

This change is necessary because the cache is no longer shared when cloning a Butler instance.
 ctrl_mpexec registers DatasetType in one instance of a cloned Butler, but does work with the DatasetType in another instance.

It will also allow the server to use DatasetTypes added by an external process without needing to restart the server.