Prefetch dimension record cache in Butler factory
In DM-42324, the dimension record cache was changed so that it is no longer shared between Butler instances.  This cache is somewhat expensive to fetch and is used for almost every Butler operation.

LabeledButlerFactory now downloads this cache into its 'template' DirectButler instances when they are first created, so it is available for copying into clones.