Do not attempt to convert the label to a ResourcePath in butler config
With the recent change in ResourcePath to force /repo/main to an
absolute URI the code that was seeing if anything was found
got confused a bit because file:///repo/main was not the same
as /repo/main. Change it to allow the get_repo_uri call
to raise and if it raises continue to use the original label.