Move find-first specification from QueryTree to result specs.
The downside of this is that the InQueryTree predicate now can't be
given a find-first subquery, but that was always a dubious proposition
for query performance, and I don't think the OODS use case that
motivated InQueryTree actually requires it.

The upside is that this is a nice simplification that just feels like
a better division of responsibilities.