Update HSC test data to new universe version, add day_obs.
This file was so old it predated the addition of day_obs to the
exposure and visit dimensions, and that was back before we even had a
dimension universe version.  While it's nice to have the YAML import
code migrate simple things, it has no idea where an instrument is and
hence can't handle this one.

Procedure was:

- make an empty SQLite repo

- import the old file

- query for visit dimensions (there are no exposures in this file)

- make new records with day_obs set, by looking at
  visit.timespan.begin and doing some timezone stuff (HSC is in
  Hawaii, which I just hard-coded here)

- re-insert the new visit recods with replace=True

- export all of the dimension elements in butler.dimensions.elements
  with element.has_own_table=True.