Avoid unnecessary DB lookups in Datastore.knows
When the DatasetRef contains datastore records, we now use them as the definitive source of truth for FileDatastore.knows().  This allows the function to avoid database I/O when datastore records are attached.

The type of DatasetDatastoreRecords was made more exact -- code is assuming the truthy/falsy behavior of the Iterable as if it is was a list, and in practice it is always a list.