Move testPruneDatasets to cover more configurations
Move testPruneDatasets up a level in the test hierarchy to FileDatastoreButlerTests, because it should work for any configuration that does not use an InMemoryDatastore.

This is also the only place where Butler.exists and Butler._exists_many are thoroughly covered, so it will be needed to get coverage for these methods in RemoteButler.