Match RemoteButler.exists behavior to DirectButler
Updated RemoteButler file info endpoints so that they can represent the DirectButler concept of "datasets known to the registry but not known to the datastore", which means roughly that we know about an artifact that used to exist but no longer does.

This allows us to make RemoteButler.exists match the DirectButler behavior, which has separate bitflags for "datastore exists" and "registry exists".