Ensure missing server error handling is not hidden
In unit tests, we now convert unhandled server errors to a special exception class 'UnhandledServerError' that is not part of Butler's usual exception hierarchy.

The FastAPI TestClient by default passes unhandled exceptions up from the server to the client.  This is useful behavior for unit testing because it gives you traceability from the test to the problem in the server code. However, because RemoteButler is in some ways just a proxy for the server-side Butler, we raise similar exceptions on the client and server side. Thus the default TestClient behavior can mask missing error-handling logic.

The registry test suite in test_remote_butler.py does not yet have all error-handling logic in place, so it opts out of the new behavior.