Use real exception handling for RemoteButler test
Previously the RemoteButler registry tests relied on exceptions bubbling up from the server to a client in a way that wouldn't work in a real deployment.  Now that an exception propagation framework is in place for all of the DirectButler code called by the server, we can remove this hack.