Make Timespan into a Pydantic model.
This requires making Timespan.nsec public, which is only slightly
unfortunate; we can declare it frozen to maintain immutability, and
we couldn't realistically change the implementation in the future given
that we've already exposed that in how it's serialized.

This obsoletes the previous commit on this branch that used
pydantic_core, but I'm keeping that version in the commit history in
case it's meaningfully more efficient in validation or memory use (I
suspect it's slightly more efficient, but not enough to matter).