Handle default data ID for RemoteQueryDriver
Implement stubs for RemoteButler's RegistryDefaults data ID handling.
Pass the default data ID from RemoteButler to the server so it can be used by the server's query driver.

One side effect of this change is that RemoteButler now calls RegistryDefaults.finish() during initialization, which triggers network traffic to download the dimension universe from the server.  Some tests that weren't expecting the server to be accessed that early needed to be tweaked as a result.