For transfers check bidirectional conversion support
A butler.get() will internally get the old python type
from the formatter but must be able to convert to the new python
type.

We also have to allow for the user to be able to get the original
python type. Butler will think it has to do a conversion from
the native type to this requested type and will fail if no
conversion exists even though it doesn't need to do a conversion.

Additionally, to avoid confusion we need to allow get to work
of the original type even for datasets that were put natively
in this butler.

This all implies that the safe approach for transfers is to
insist that both conversions exist.