Create an interface for converting rows to pages
Implementing DatasetRef result type requires some data that QueryDriver has but which would be awkward to pass through _Cursor.  _Cursor is also just starting to have a lot of arguments to its constructor.

To fix this (as well as separate concerns a little bit), define an interface for converting SQL rows into ResultPages and have QueryDriver inject a dependency into _Cursor for doing that work.