Make ParentDatasetQueryResults an interface
Made ParentDatasetQueryResults an abstract base class, with a new concrete class DatabaseParentDatasetQueryResults implementing it.

There is a weird circular dependency between ParentDatasetQueryResults and its base class, DatasetQueryResults.  DatasetQueryResults defines functions that return a ParentDatasetQueryResults.

Thus to be able to create a new implementation of DatasetQueryResults, the new implementation has to be a ParentDatasetQueryResults.