Change approach to kwargs merging
Using Config is fine when merging dicts but it breaks if
a top-level kwarg is a Mapping but should not be merged.
This was being triggered when formatter constructors
switched to using dataId=DataCoordinate. Reorganize
the merging such that only keys that appear in both
kwargs are merged using Config.