Fix a sqlalchemy crash when no datasets found
Fixed an issue where sqlalchemy would fail to compile the select statement for dataset queries in certain edge cases where we were ending up with empty collection lists.  This was happening because of an optimization that tries to avoid a join by generating a case statement.  With no collections there were no cases for the case statement, and sqlalchemy was crashing out.