Fix server-side cursor lifetime issues (DM-43697)
Server side cursors need extra care when managing results, if the result
is not exhausted then cursor stays open. This caused error when deleting
temporary tables with cursors open against those tables. The easy way to
close the cursor is to use result as a context manager - cursors are
guaranteed to be closed on exit from a context.