Skip datastore root check for non-POSIX filesystems
We now skip existence checks for the datastore root for non-POSIX root paths.  This fixes an issue where FileDatastore was throwing a permissions-related exception on startup when trying to use a read-only S3 bucket via either `s3://` or `https://`.

This was occurring because on S3 there is no concept of a directory, so even though files exist in the datastore the root directory does not.