Fix type issue found by mypy 1.11
We use a default_key=None in the log handling code to imply
the root logger. Changing the types in split_kv to use str | None
fixes the problem. The --log-levels option wants to use None
to indicate the default logger should be affected and saves
"" to mean the root logger (the default logger is "lsst").