Disable default acceptance on put
ci_middleware has mocks that look like they are all compatible
with the requested python type when they aren't so by default
a formatter can't naively compare types and decide that since
they match no conversion is required. Always return False for
Formatter.can_accept and force type coercion (which works
with the mocks because of subclassing). Formatter can't know
about mocks (which are in pipe_base) and returning False
doesn't change anything in the default case because conversion
will be a no-op.

The mocks also track file conversions so if conversions are
bypassed the tests confirming conversions won't pass.