Add a way for Butler server to serve static files
You can now set the environment variable DAF_BUTLER_SERVER_STATIC_FILES_PATH to point to a directory of static files to be served by Butler server from its `/configs/` HTTP path.

There isn't currently a good place for the DirectButler configuration files used in the RSP lab containers to live.  These need to be controlled by Phalanx so we can update them as part of the IDF deployment process, but Phalanx doesn't provide a convenient way of deploying static files.  Much of this configuration is shared with Butler server or updated at the same time that Butler server is updated.  So it is convenient for the moment to just serve these files from Butler server rather than standing up additional infrastructure to host them.