Add unit test for general queries and fix serialization of ingest_date.
Adding unit tests exposed the issue with serialization of ingest_time.
We still create schema with a native time type for that column, while
its corresponding ColumnSpec says it should be astropy time. The code
that I added recently used SerializableTime with type adapter to
serialize it which did not work with datetime. I had to reimplement
serialization methods to allow more flexible handling of ingest_date types.