Return ingest_date as astropy time from general query (DM-45764)
There is no easy way to catch/convert ingest_date in the rows returned
from all queries. For now only the general query can return ingest_date,
so it is an easy fix for that case, though it requires checking type
of the value in that column for each row.