Revert "Fix --show-uri for chained datasstores"
This reverts commit a287d0bd7ebe3a989f298acda666d07442d735da.

After some thought it became clear that get_many_uris guarantees
only a single URI per dataset ref regardless of whether a
chained datastore is present so this additional complexity is
not needed. Tables are created for each component dataset type
so even for composite disassembly there will not be multiple
URIs in a table for the same dataset ref.