Fix summary_datasets being ignored
When attempting to add a test for the summary_datasets parameter to ButlerCollectionInfo for RemoteButler, it turned out that it does not doing anything in the DirectButler version.

This was occurring because a caching context causes this parameter to be ignored, and the cache was always enabled in query_info previously.  This cache does not have a benefit for the new implementation.