Fix issue with "collection" dataset field
Fix an issue similar to the previous commit, where query_datasets would fail on Postgres 16 with the error 'psycopg2.errors.DatatypeMismatch: could not determine polymorphic type because input has type unknown'.

This was occurring when there was a single collection specified for the dataset search, causing a literal  "collection" dataset field value to be included in an any_value aggregate function.