Allow queries with OR or IN constraints on governor dimensions.
Instead of requiring a new-query-system WHERE constraint to lock down
the (e.g.) instrument to a single value in order to allow a dependent
value (e.g. detector) to appear, we now just check that instrument
appears somewhere in the expression, since that'd catch the most likely
mistake of leaving the instrument out entirely (or assuming incorrectly
that a default would be available).